backend/intelligence/requirements.txt
smolagents
python-dotenv
protobuf
sentencepiece
llama-cpp-python
smolagents[litellm]